chore: remove ProgramDriver and Driver classes
