+++ src/types/HasDriver.ts
@@ -1,0 +1,4 @@
+export type HasDriver<T> = {
+  driver: () => T;
+  setDriver: (newDriver: T) => void;
+};
