+++ src/types/index.ts
@@ -2,13 +2,12 @@ export * from './Amount';
 export * from './BaseAccount';
 export * from './Cluster';
 export * from './DateTimeString';
-export * from './Driver';
+export * from './HasDriver';
 export * from './HasMetaplex';
 export * from './MetaplexPlugin';
 export * from './Model';
 export * from './Operation';
 export * from './Pda';
 export * from './Program';
-export * from './ProgramDriver';
 export * from './PublicKeyString';
 export * from './Signer';
